Rework tests and remove debug statement
